available in dev env
https://app73204e59.azurewebsites.net/

### How to create and deploy a backend model?

Look in the `/ml` directory for a notebook

```sh
cd ml/
conda activate signalboxml # if already created
pip install -t requirements.txt
jupyter notebook

```

- update ML code to match recommender
- refer label to name
- update notebook for latest recommender version
- update client definitions
- improve exceptions
- ensure common dynamic serialisation
- inputs and outputs for ML models #3620
- recommenders as models v1
- update src of entry.py (from notebook)
- can create and test recommenders #3633
- Improve UI of parameter set recommenders
- update signalbox.js
- update migrations

Related work items: #3620, #3633

+++ app/core/models/events/MomentCount.cs
@@ -4,13 +4,13 @@ namespace SignalBox.Core
 {
     public class MomentCount
     {
-        public MomentCount(DateTimeOffset timestamp, int count)
+        public MomentCount(DateTimeOffset timestamp, long count)
         {
             Timestamp = timestamp;
             Count = count;
         }
         
-        public MomentCount(string category, DateTimeOffset timestamp, int count)
+        public MomentCount(string category, DateTimeOffset timestamp, long count)
         {
             Category = category;
             Timestamp = timestamp;
@@ -20,6 +20,6 @@ namespace SignalBox.Core
         public string Category { get; set; }
         public DateTimeOffset Timestamp { get; set; }
         public long UnixTime => Timestamp.ToUnixTimeMilliseconds();
-        public int Count { get; set; }
+        public long Count { get; set; }
     }
 }
